+++ CHANGELOG.md
@@ -4,6 +4,8 @@ on [Keep a CHANGELOG](http://keepachangelog.com/). This project adheres to
 [Semantic Versioning](http://semver.org/).
 
 [Unreleased]
+### Fixed
+- Fixed MeterNotFoundException thrown when looking up metrics meter with unknown tags
 
 # [17.103.0-M14] - 2025-06-24
 ### Changed